pyproject.toml
[project]
name = "day17-2a202600417-phamminhkhang"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.8",
    "langgraph>=1.1.9",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "tiktoken>=0.12.0",
]
